Port main thread to typescript
