+++ src/main/index.d.ts
@@ -1,9 +1,14 @@
 import type { IProductInformation } from '../shared'
 import type { BrowserWindow } from 'electron'
-declare global
-{
-    var electronMainWindow: BrowserWindow
-    var __SIXGRID_PRODUCT_BUILD_VERSION: string
-    var __PRODUCT_EXTENDED_INFORMATION: IProductInformation
+declare global {
+    namespace NodeJS {
+        interface Global {
+            electronMainWindow: BrowserWindow|null
+            __SIXGRID_PRODUCT_BUILD_VERSION: string
+            __PRODUCT_EXTENDED_INFORMATION: IProductInformation
+            __static: string
+            debugMode: boolean
+        }
+    }
 }
 export {}
