+++ src/main/menu.ts
@@ -1,6 +1,7 @@
 import * as helpers from './helpers'
+import {MenuItemConstructorOptions, MenuItem} from 'electron'
 const isMac = process.platform == 'darwin'
-export default [
+const menuTemplate: any = [
     ...(isMac ? [{ role: 'appMenu' }] : []),
     {
         label: 'File',
@@ -36,10 +37,11 @@ export default [
                 label: 'Toggle Developer Tools',
                 accelerator: 'F12',
                 click: () => {
-                    if (electronMainWindow != undefined && electronMainWindow.mainWindow != undefined)
-                        electronMainWindow.webContents.openDevTools()
+                    if (global.electronMainWindow != undefined && global.electronMainWindow.webContents != undefined)
+                        global.electronMainWindow.webContents.openDevTools()
                 }
             }
         ]
     }
 ]
+export default menuTemplate
